* Move Linking to tanka.server.links assembly (move tests to .tests) * Disable digitransit tests as the digitransit server is returning 500 error
