+++ src/graphql.server.links/HubConnectionExtensions.cs
@@ -4,18 +4,18 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.SignalR.Client;
 using Tanka.GraphQL.DTOs;
 
-namespace Tanka.GraphQL.Linking
+namespace Tanka.GraphQL.Server.Links
 {
     public static class HubConnectionExtensions
     {
         public static async Task<ChannelReader<ExecutionResult>> StreamQueryAsync(
-            this HubConnection connection, 
+            this HubConnection connection,
             QueryRequest query,
             CancellationToken cancellationToken)
         {
             var channel = await connection.StreamAsChannelAsync<ExecutionResult>(
                 "query",
-                query, 
+                query,
                 cancellationToken);
 
             return channel;
