+++ src/graphql.server.links/RemoteLinks.cs
@@ -7,10 +7,10 @@ using System.Threading.Channels;
 using System.Threading.Tasks;
 using GraphQLParser.AST;
 using Microsoft.AspNetCore.SignalR.Client;
-using Tanka.GraphQL.Language;
 using Tanka.GraphQL.DTOs;
+using Tanka.GraphQL.Language;
 
-namespace Tanka.GraphQL.Linking
+namespace Tanka.GraphQL.Server.Links
 {
     public static class RemoteLinks
     {
