+++ src/graphql.server.links/graphql.server.links.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <AssemblyName>tanka.graphql.server.links</AssemblyName>
+    <RootNamespace>Tanka.GraphQL.Server.Links</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\graphql\graphql.csproj" />
+  </ItemGroup>
+
+</Project>
