+++ tests/graphql.server.links.tests/MakeRemoteExecutableDigitransitFacts.cs
@@ -6,14 +6,11 @@ using System.Text;
 using System.Threading.Channels;
 using System.Threading.Tasks;
 using Tanka.GraphQL.Introspection;
-using Tanka.GraphQL.Linking;
 using Tanka.GraphQL.SchemaBuilding;
-using Tanka.GraphQL.Tests.Data;
-using Tanka.GraphQL.Tools;
 using Tanka.GraphQL.TypeSystem.ValueSerialization;
 using Xunit;
 
-namespace Tanka.GraphQL.Tests.Tools
+namespace Tanka.GraphQL.Server.Links.Tests
 {
     public class MakeRemoteExecutableDigitransitFacts
     {
@@ -26,7 +23,7 @@ namespace Tanka.GraphQL.Tests.Tools
             return reader.ReadToEnd();
         }
 
-        [Fact]
+        [Fact(Skip = "Digitransit returning 500 error")]
         public async Task ExecuteRemotely()
         {
             /* Given */
@@ -73,7 +70,7 @@ namespace Tanka.GraphQL.Tests.Tools
             ");
         }
 
-        [Fact]
+        [Fact(Skip = "Digitransit returning 500 error")]
         public async Task ExecuteRemotely_with_link_graphql_error()
         {
             /* Given */
@@ -150,7 +147,7 @@ namespace Tanka.GraphQL.Tests.Tools
             ");
         }
 
-        [Fact]
+        [Fact(Skip = "Digitransit returning 500 error")]
         public async Task ExecuteWithStaticDataLink()
         {
             /* Given */
@@ -201,7 +198,7 @@ namespace Tanka.GraphQL.Tests.Tools
             ");
         }
 
-        [Fact]
+        [Fact(Skip = "Digitransit returning 500 error")]
         public async Task RemoteExecute_with_link_exception()
         {
             /* Given */
