+++ tests/graphql.server.links.tests/graphql.server.links.tests.csproj
@@ -1,0 +1,25 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
+
+    <IsPackable>false</IsPackable>
+
+    <AssemblyName>tanka.graphql.server.links.tests</AssemblyName>
+
+    <RootNamespace>Tanka.GraphQL.Server.Links.Tests</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.2.0" />
+    <PackageReference Include="xunit" Version="2.4.0" />
+    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.0" />
+    <PackageReference Include="coverlet.collector" Version="1.0.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\graphql.server.links\graphql.server.links.csproj" />
+    <ProjectReference Include="..\..\src\graphql\graphql.csproj" />
+  </ItemGroup>
+
+</Project>
